BUGFIX: Fix compiler warning: comparison between signed and unsigned integer expressions
git-svn-id: https://assimp.svn.sourceforge.net/svnroot/assimp/trunk@520 67173fc5-114c-0410-ac8e-9d2fd5bffc1f

+++ code/3DSLoader.cpp
@@ -1040,7 +1040,7 @@ void Discreet3DSImporter::ParseMeshChunk()
 
 		// Larger 3DS files could have multiple FACE chunks here
 		chunkSize = stream->GetRemainingSizeToLimit();
-		if (chunkSize > sizeof(Discreet3DS::Chunk))
+		if ( chunkSize > (int) sizeof(Discreet3DS::Chunk ) )
 			ParseFaceChunk();
 		}
 		break;
